:hammer: refector: Calendar 리펙토링
네이밍 수정 및 constant.js 정리
