+++ src/components/Diary/DiaryMainContainer/sections/CenterSection/Calender/DateBox.scss
@@ -16,7 +16,8 @@
     text-align: center;
     border: 0 !important;
   }
-  .hasData {
+
+  .hasDataIcon {
     position: absolute;
     border-radius: 50%;
     background-color: #d9d9d9;
